requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
asgiref==3.11.0
attrs==25.3.0
Automat==24.8.1
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudscraper==1.2.71
constantly==23.10.4
coverage==7.12.0
cryptography==44.0.2
cssselect==1.3.0
defusedxml==0.7.1
Deprecated==1.2.18
Django==5.2.8
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-yasg==1.21.11
filelock==3.18.0
flasgger==0.9.7.1
Flask==3.1.2
Flask-Limiter==3.12
frozenlist==1.8.0
h11==0.16.0
hltv_async_api==0.8.3
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
iniconfig==2.3.0
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
limits==4.4.1
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mistune==3.1.4
multidict==6.7.0
numpy==2.3.5
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.3.3
parsel==1.10.0
pluggy==1.6.0
propcache==0.4.1
Protego==0.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
Pygments==2.19.1
pyOpenSSL==25.0.0
pyparsing==3.2.5
PySocks==1.7.1
pytest==9.0.1
pytest-cov==7.0.0
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
queuelib==1.7.0
referencing==0.37.0
requests==2.32.5
requests-file==2.1.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.30.0
Scrapy==2.13.4
scrapy-selenium==0.0.7
selenium==4.30.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
sqlparse==0.5.4
TableauScraper==0.1.29
tldextract==5.1.3
trio==0.29.0
trio-websocket==0.12.2
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
w3lib==2.3.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
yarl==1.22.0
zope.interface==7.2
